pyproject.toml
[project]
name = "langsmith-skills"
version = "0.1.0"
description = "LangSmith skills for Claude Code - trace querying and dataset generation"
requires-python = ">=3.11"
dependencies = [
    "langsmith>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "deepagents>=0.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pexpect>=4.9.0",
]

[tool.setuptools]
packages = []
